+++ packages/lumx-react/src/utils/type.ts
@@ -59,6 +59,15 @@ export interface HasClassName {
     className?: string;
 }
 
+
+export interface HasCloseMode {
+    /**
+     * Choose how the children are hidden when closed
+     * ('hide' keeps the children in DOM but hide them, 'unmount' remove the children from the DOM).
+     */
+    closeMode?: 'hide' | 'unmount';
+}
+
 /**
  * Define a generic props types.
  */