+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   `Slideshow`: changed active pagination item width for better a11y.
 -   `ImageLightbox`: fix closing animation cut short because of unstable image reference.
 
+### Added
+
+- `Tooltip`: Add `closeMode` to hide the tooltip instead of unmounting it
+
 ## [3.9.1][] - 2024-09-17
 
 ### Fixed